Düzeltmeler
TCPClient da dosya yolu düzeltildi,bağlantı yoksa ekranı kilitle metodu
class'a taşındı.

+++ ResOtomasyon/src/main/java/com/res_otomasyon/resotomasyon/MyExpandableListAdapter.java
@@ -328,7 +328,7 @@ public class MyExpandableListAdapter extends BaseExpandableListAdapter {
                                     int adet = 1;
 
                                     if (urunBilgileri.length > 1)
-                                        adet = Integer.parseInt(urunBilgileri[1]);
+                                        adet = (int)(Double.parseDouble(urunBilgileri[1]));
 
                                     adet--;
